add : httpmanager 파일 분리
