+++ drawingnote_application/lib/drawingpainter.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+
+// Painter 클래스
+class DrawingPainter extends CustomPainter {
+  final List<List<Offset?>> lines;
+
+  final Offset offset;
+
+  DrawingPainter(this.lines, {this.offset = Offset.zero});
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    Paint paint = Paint()
+      ..color = Colors.black
+      ..strokeCap = StrokeCap.round
+      ..strokeWidth = 5.0;
+
+    for (var line in lines) {
+      for (int i = 0; i < line.length - 1; i++) {
+        if (line[i] != null && line[i + 1] != null) {
+          canvas.drawLine(line[i]!, line[i + 1]!, paint);
+        }
+      }
+    }
+  }
+
+  @override
+  bool shouldRepaint(covariant CustomPainter oldDelegate) {
+    return true;
+  }
+}
